Extract model relationships into single use traits.
No functional changes, just moves the relationships themselves into
files in the App/Models/Relationships directory, to clean up the model
files themselves a bit.
