+++ app/Models/AssetMaintenance.php
@@ -2,6 +2,7 @@
 namespace App\Models;
 
 use App\Helpers\Helper;
+use App\Models\Relationships\AssetMaintenanceRelationships;
 use App\Models\Traits\Searchable;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Lang;
@@ -18,7 +19,7 @@ class AssetMaintenance extends Model implements ICompanyableChild
     use SoftDeletes;
     use CompanyableChildTrait;
     use ValidatingTrait;
-
+    use AssetMaintenanceRelationships;
 
     protected $dates = [ 'deleted_at', 'start_date' , 'completion_date'];
     protected $table = 'asset_maintenances';
@@ -115,42 +116,6 @@ class AssetMaintenance extends Model implements ICompanyableChild
             $value = null;
         }
         $this->attributes['completion_date'] = $value;
-    }
-
-    /**
-       * asset
-       * Get asset for this improvement
-       *
-       * @return mixed
-       * @author  Vincent Sposato <vincent.sposato@gmail.com>
-       * @version v1.0
-       */
-    public function asset()
-    {
-
-        return $this->belongsTo('\App\Models\Asset', 'asset_id')
-                    ->withTrashed();
-    }
-
-    /**
-     * Get the admin who created the maintenance
-     *
-     * @return mixed
-     * @author  A. Gianotto <snipe@snipe.net>
-     * @version v3.0
-     */
-    public function admin()
-    {
-
-        return $this->belongsTo('\App\Models\User', 'user_id')
-            ->withTrashed();
-    }
-
-    public function supplier()
-    {
-
-        return $this->belongsTo('\App\Models\Supplier', 'supplier_id')
-                    ->withTrashed();
     }
 
     /**
