+++ app/Models/AssetModel.php
@@ -1,6 +1,7 @@
 <?php
 namespace App\Models;
 
+use App\Models\Relationships\AssetModelRelationships;
 use App\Models\Requestable;
 use App\Models\SnipeModel;
 use App\Models\Traits\Searchable;
@@ -19,7 +20,7 @@ class AssetModel extends SnipeModel
 {
     use SoftDeletes;
     protected $presenter = 'App\Presenters\AssetModelPresenter';
-    use Requestable, Presentable;
+    use Requestable, Presentable, AssetModelRelationships;
     protected $dates = ['deleted_at'];
     protected $table = 'models';
     protected $hidden = ['user_id','deleted_at'];
@@ -88,41 +89,6 @@ class AssetModel extends SnipeModel
         'category'     => ['name'],
         'manufacturer' => ['name'],
     ];      
-
-    public function assets()
-    {
-        return $this->hasMany('\App\Models\Asset', 'model_id');
-    }
-
-    public function category()
-    {
-        return $this->belongsTo('\App\Models\Category', 'category_id');
-    }
-
-    public function depreciation()
-    {
-        return $this->belongsTo('\App\Models\Depreciation', 'depreciation_id');
-    }
-
-    public function adminuser()
-    {
-        return $this->belongsTo('\App\Models\User', 'user_id');
-    }
-
-    public function manufacturer()
-    {
-        return $this->belongsTo('\App\Models\Manufacturer', 'manufacturer_id');
-    }
-
-    public function fieldset()
-    {
-        return $this->belongsTo('\App\Models\CustomFieldset', 'fieldset_id');
-    }
-
-    public function defaultValues()
-    {
-        return $this->belongsToMany('\App\Models\CustomField', 'models_custom_fields')->withPivot('default_value');
-    }
 
 
     public function getImageUrl() {
