+++ app/Models/Department.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use App\Http\Traits\UniqueUndeletedTrait;
+use App\Models\Relationships\DepartmentRelationships;
 use App\Models\Traits\Searchable;
 use Illuminate\Database\Eloquent\Model;
 use Log;
@@ -22,7 +23,7 @@ class Department extends SnipeModel
      */
     protected $injectUniqueIdentifier = true;
 
-    use ValidatingTrait, UniqueUndeletedTrait;
+    use ValidatingTrait, UniqueUndeletedTrait, DepartmentRelationships;
 
     protected $rules = [
         'name'            => 'required|max:255',
@@ -63,36 +64,6 @@ class Department extends SnipeModel
     protected $searchableRelations = [];    
 
 
-    public function company()
-    {
-        return $this->belongsTo('\App\Models\Company', 'company_id');
-    }
-
-    /**
-     * Even though we allow allow for checkout to things beyond users
-     * this method is an easy way of seeing if we are checked out to a user.
-     * @return mixed
-     */
-    public function users()
-    {
-        return $this->hasMany('\App\Models\User', 'department_id');
-    }
-
-
-    /**
-    * Return the manager in charge of the dept
-    * @return mixed
-    */
-    public function manager()
-    {
-        return $this->belongsTo('\App\Models\User', 'manager_id');
-    }
-
-
-    public function location()
-    {
-        return $this->belongsTo('\App\Models\Location', 'location_id');
-    }
     
     /**
      * Query builder scope to order on location name
